+++ docs/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 # unreleased - YYYY-MM-DD
 ## Python API
-* [Added] Automatically decompress gzip'ed package entries when deserializing.
+* [Added] Automatically decompress gzip'ed package entries when deserializing ([#2677](https://github.com/quiltdata/quilt/pull/2677))
 
 # 4.0.0 - 2022-01-31
 ## Python API