+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 # unreleased - YYYY-MM-DD
 ## Python API
 * [Added] Automatically decompress gzip'ed package entries when deserializing ([#2677](https://github.com/quiltdata/quilt/pull/2677))
+* [Added] Semi-atomic push ([#2689](https://github.com/quiltdata/quilt/pull/2689))
 
 # 4.0.0 - 2022-01-31
 ## Python API