+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@
 ## Catalog, Lambdas
 !-->
 
-# unreleased - YYYY-MM-DD
+# 4.1.0 - 2022-02-22
 ## Python API
 * [Added] Automatically decompress gzip'ed package entries when deserializing ([#2677](https://github.com/quiltdata/quilt/pull/2677))
 * [Added] Semi-atomic push ([#2689](https://github.com/quiltdata/quilt/pull/2689))